Update atualizado de todas as Entitys, CRUD funcionando completamente

+++ src/main/java/com/ayvy/api_java/infrastructure/entitys/Produto.java
@@ -27,7 +27,7 @@ public class Produto {
     private String nome_produto;
 
     @Column(name = "preco")
-    private Integer preco;
+    private Float preco;
 
     //Faltam:
     //Descrição String
